reorganized test classes to separate out different test plugins
