+++ data-prepper-plugin-framework/src/test/java/org/opensearch/dataprepper/plugins/configtest/TestDISourceWithConfig.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.test;
+package org.opensearch.dataprepper.plugins.configtest;
 
 import org.opensearch.dataprepper.model.annotations.DataPrepperPlugin;
 import org.opensearch.dataprepper.model.annotations.DataPrepperPluginConstructor;
